20563 license list (#53)
* Changed Look Of Welcome Page

To fix this, I also made some changes to the css structure.

Ref: #20562

* Changed About And Pricacy To Look As Welcome

The header is adjusted if the user is logged in or not. That should affect the main part of the page as well.

Ref: #20562

* Added scss files

These files hold the formatting. They are divided per component mainly to not get too big files.

Ref: #20563

* Cleaned Up Html

Adjusted to have cleaner html with css classes.

Ref: #20563

* Added New Resource Tile

The resource tiles will be shown in the resource list.

Ref: #20563

* Improved List of Resources

Changed to make it look cleaner and also changed the resource details to be a modal.

Ref: #20563

* Fixed WCAG Issues

There were some issues with regards to WCAG that are now fixed. They included contrast issues and similar links being nearby each other.

Ref: #20563

---------

Co-authored-by: Henrik Jademyr <Henrik.Jademyr@goteborgsregionen.se>